requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.3
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-phonenumber-field==6.1.0
django-ranged-response==0.2.0
django-simple-captcha==0.5.17
phonenumbers==8.12.45
Pillow==9.0.1
psycopg2==2.9.3
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
djangorestframework~=3.13.1